+++ src/lungo_cli/plugins/stirlingpdf/plugin.py
@@ -22,9 +22,9 @@ class Plugin(BasePlugin[BaseSettings]):
         return self.storage.cache_latest_dir / self.manifest.name / "logs"
 
     @override
-    def get_render_context(self) -> dict[str, Any]:
+    def get_custom_rendering_context(self) -> dict[str, Any]:
         return {"stirlingpdf_version": "0.22.8"}
 
     @override
-    def update_data(self) -> None:
+    def on_plugin_initialization(self) -> None:
         self.file_utils.create_dir(self.logs_dir)