+++ src/lungo_cli/plugins/xray/plugin.py
@@ -37,7 +37,7 @@ class Plugin(BasePlugin[Settings]):
         return Settings
 
     @override
-    def get_render_context(self) -> dict[str, Any]:
+    def get_custom_rendering_context(self) -> dict[str, Any]:
         salt = UUID(self.file_utils.read_text(self.salt_file))
 
         return {
@@ -49,7 +49,7 @@ class Plugin(BasePlugin[Settings]):
         }
 
     @override
-    def update_data(self) -> None:
+    def on_plugin_initialization(self) -> None:
         if not self.salt_file.is_file():
             self.console.print_info("Generating Xray salt...")
             self.file_utils.write_text(self.salt_file, str(uuid1()), True)